Ensure bend slur on tied bend notes.
