+++ src/rendering/glyphs/ScoreHelperNotesBaseGlyph.ts
@@ -4,11 +4,18 @@ import { BendNoteHeadGroupGlyph } from '@src/rendering/glyphs/BendNoteHeadGroupG
 import { Glyph } from '@src/rendering/glyphs/Glyph';
 import { ScoreBarRenderer } from '@src/rendering/ScoreBarRenderer';
 import { BeamDirection } from '@src/rendering/utils/BeamDirection';
+import { TieGlyph } from './TieGlyph';
 
 export class ScoreHelperNotesBaseGlyph extends Glyph {
-    private static readonly SlurHeight: number = 11;
     public static readonly EndPadding: number = ((10 / 2) | 0) + 3;
     protected BendNoteHeads: BendNoteHeadGroupGlyph[] = [];
+
+    public doLayout(): void {
+        for (const head of this.BendNoteHeads) {
+            head.renderer = this.renderer;
+            head.doLayout();
+        }
+    }
 
     protected drawBendSlur(
         canvas: ICanvas,
@@ -20,46 +27,7 @@ export class ScoreHelperNotesBaseGlyph extends Glyph {
         scale: number,
         slurText?: string
     ): void {
-        let normalVectorX: number = y2 - y1;
-        let normalVectorY: number = x2 - x1;
-        let length: number = Math.sqrt(normalVectorX * normalVectorX + normalVectorY * normalVectorY);
-        if (down) {
-            normalVectorX *= -1;
-        } else {
-            normalVectorY *= -1;
-        }
-        // make to unit vector
-        normalVectorX /= length;
-        normalVectorY /= length;
-        // center of connection
-        // TODO: should be 1/3
-        let centerX: number = (x2 + x1) / 2;
-        let centerY: number = (y2 + y1) / 2;
-        let offset: number = ScoreHelperNotesBaseGlyph.SlurHeight * scale;
-        if (x2 - x1 < 20) {
-            offset /= 2;
-        }
-        let cp1X: number = centerX + offset * normalVectorX;
-        let cp1Y: number = centerY + offset * normalVectorY;
-        canvas.beginPath();
-        canvas.moveTo(x1, y1);
-        canvas.lineTo(cp1X, cp1Y);
-        canvas.lineTo(x2, y2);
-        canvas.stroke();
-        if (slurText) {
-            let w: number = canvas.measureText(slurText);
-            let textOffset: number = down ? 0 : -canvas.font.size;
-            canvas.fillText(slurText, cp1X - w / 2, cp1Y + textOffset);
-        }
-    }
-
-    public doLayout(): void {
-        super.doLayout();
-        this.width = 0;
-        for (let noteHeads of this.BendNoteHeads) {
-            noteHeads.doLayout();
-            this.width += noteHeads.width + 10 * this.scale;
-        }
+        TieGlyph.drawBendSlur(canvas, x1, y1, x2, y2, down, scale, slurText);
     }
 
     protected getBeamDirection(beat: Beat, noteRenderer: ScoreBarRenderer): BeamDirection {
