+++ src/rendering/glyphs/TieGlyph.ts
@@ -78,19 +78,37 @@ export class TieGlyph extends Glyph {
             shouldDraw = true;
         }
         if (shouldDraw) {
-            TieGlyph.paintTie(
-                canvas,
-                this.scale,
-                startX,
-                startY,
-                endX,
-                endY,
-                this.tieDirection === BeamDirection.Down,
-                this.getTieHeight(startX, startY, endX, endY),
-                4
-            );
+            if(this.shouldDrawBendSlur()) {
+                TieGlyph.drawBendSlur(
+                    canvas,
+                    startX,
+                    startY,
+                    endX,
+                    endY,
+                    this.tieDirection === BeamDirection.Down,
+                    this.scale
+                );
+            }
+            else {
+                TieGlyph.paintTie(
+                    canvas,
+                    this.scale,
+                    startX,
+                    startY,
+                    endX,
+                    endY,
+                    this.tieDirection === BeamDirection.Down,
+                    this.getTieHeight(startX, startY, endX, endY),
+                    4
+                );
+            }
+            
             canvas.fill();
         }
+    }
+
+    protected shouldDrawBendSlur() {
+        return false;
     }
 
     protected getTieHeight(startX: number, startY: number, endX: number, endY: number): number {
@@ -172,5 +190,50 @@ export class TieGlyph extends Glyph {
         canvas.quadraticCurveTo(cp1X, cp1Y, x2, y2);
         canvas.quadraticCurveTo(cp2X, cp2Y, x1, y1);
         canvas.closePath();
+    }
+
+    private static readonly BendSlurHeight: number = 11;
+
+    public static drawBendSlur(
+        canvas: ICanvas,
+        x1: number,
+        y1: number,
+        x2: number,
+        y2: number,
+        down: boolean,
+        scale: number,
+        slurText?: string
+    ): void {
+        let normalVectorX: number = y2 - y1;
+        let normalVectorY: number = x2 - x1;
+        let length: number = Math.sqrt(normalVectorX * normalVectorX + normalVectorY * normalVectorY);
+        if (down) {
+            normalVectorX *= -1;
+        } else {
+            normalVectorY *= -1;
+        }
+        // make to unit vector
+        normalVectorX /= length;
+        normalVectorY /= length;
+        // center of connection
+        // TODO: should be 1/3
+        let centerX: number = (x2 + x1) / 2;
+        let centerY: number = (y2 + y1) / 2;
+        let offset: number = TieGlyph.BendSlurHeight * scale;
+        if (x2 - x1 < 20) {
+            offset /= 2;
+        }
+        let cp1X: number = centerX + offset * normalVectorX;
+        let cp1Y: number = centerY + offset * normalVectorY;
+        canvas.beginPath();
+        canvas.moveTo(x1, y1);
+        canvas.lineTo(cp1X, cp1Y);
+        canvas.lineTo(x2, y2);
+        canvas.stroke();
+        if (slurText) {
+            let w: number = canvas.measureText(slurText);
+            let textOffset: number = down ? 0 : -canvas.font.size;
+            canvas.fillText(slurText, cp1X - w / 2, cp1Y + textOffset);
+        }
     }
 }
